Fix loading of image assets
Image paths must be available without digest fingerprints because
the Leaflet CSS refers to them by relative path, and the Ember
CSS preprocessing will NOT rewrite the image URLs to include the
digests. To do this a Ruby gem (non-stupid-digest-assets) was
added.

Also because the CSS assumes the images are in a relative directory
of 'images', we end up with the redundant 'assets/images/images'.